pyproject.toml
[tool.poetry]
name = "post-creation-service"
version = "0.1.0"
description = ""
authors = ["Denis Tokarev <wf.denis.tokarev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
asyncpraw = "^7.8.0"
tweepy = {extras = ["async"], version = "^4.14.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
